+++ tests/test_api.py
@@ -4,6 +4,7 @@ send coroutine testing.
 
 import asyncio
 import email
+import email.message
 import pathlib
 import socket
 import ssl
@@ -159,14 +160,14 @@ async def test_send_with_login(
 async def test_send_via_socket(
     hostname: str,
     smtpd_server_port: int,
-    message: email.message.Message,
+    mime_message: email.message.EmailMessage,
     received_messages: list[email.message.EmailMessage],
 ) -> None:
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
         sock.connect((hostname, smtpd_server_port))
 
-        errors, response = await send(
-            message,
+        errors, _ = await send(
+            mime_message,
             hostname=None,
             port=None,
             sock=sock,
@@ -176,21 +177,65 @@ async def test_send_via_socket(
         assert not errors
         assert len(received_messages) == 1
 
-        assert sock.fileno() > 0, "Socket unexpectedly closed"
+
+@pytest.mark.smtpd_options(tls=True)
+async def test_send_via_socket_tls_and_hostname(
+    hostname: str,
+    client_tls_context: ssl.SSLContext,
+    smtpd_server_port: int,
+    mime_message: email.message.EmailMessage,
+    received_messages: list[email.message.EmailMessage],
+) -> None:
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
+        sock.connect((hostname, smtpd_server_port))
+
+        errors, _ = await send(
+            mime_message,
+            hostname=hostname,
+            port=None,
+            sock=sock,
+            tls_context=client_tls_context,
+            use_tls=True,
+        )
+
+        assert not errors
+        assert len(received_messages) == 1
 
 
 async def test_send_via_socket_path(
     smtpd_server_socket_path: asyncio.AbstractServer,
     socket_path: Union[pathlib.Path, str, bytes],
-    message: email.message.Message,
+    mime_message: email.message.EmailMessage,
     received_messages: list[email.message.EmailMessage],
 ) -> None:
-    errors, response = await send(
-        message,
+    errors, _ = await send(
+        mime_message,
         hostname=None,
         port=None,
         socket_path=socket_path,
         start_tls=False,
+    )
+
+    assert not errors
+    assert len(received_messages) == 1
+
+
+@pytest.mark.smtpd_options(tls=True)
+async def test_send_via_socket_path_with_tls(
+    smtpd_server_socket_path: asyncio.AbstractServer,
+    socket_path: Union[pathlib.Path, str, bytes],
+    hostname: str,
+    client_tls_context: ssl.SSLContext,
+    mime_message: email.message.EmailMessage,
+    received_messages: list[email.message.EmailMessage],
+) -> None:
+    errors, _ = await send(
+        mime_message,
+        hostname=hostname,
+        port=None,
+        socket_path=socket_path,
+        use_tls=True,
+        tls_context=client_tls_context,
     )
 
     assert not errors